Add strict compile warnings

+++ src/brf_shift_queue.c
@@ -36,9 +36,9 @@ size_t brf_shift_queue_push(brf_shift_queue_t *q, const uint8_t *src,
   unsigned char const *buf_ptr = (unsigned char const *)src;
 
   // Get write offset for accessing queue buffer
-  const unsigned write_offset = q->write_idx & (q->capacity - 1);
+  const size_t write_offset = q->write_idx & (q->capacity - 1);
   // Find the number of bytes can write before wrap around
-  const unsigned pre_wrap_copy_len = MIN(size, q->capacity - write_offset);
+  const size_t pre_wrap_copy_len = MIN(size, q->capacity - write_offset);
   memcpy(&q->buff[write_offset], buf_ptr, pre_wrap_copy_len);
   buf_ptr += pre_wrap_copy_len;
 
@@ -49,7 +49,7 @@ size_t brf_shift_queue_push(brf_shift_queue_t *q, const uint8_t *src,
   q->write_idx += size;
 
   // Determine if read_idx needs to be bumpped due to over writting
-  const unsigned overwrite_len =
+  const size_t overwrite_len =
       MAX((q->write_idx - q->read_idx), q->capacity) - q->capacity;
   q->read_idx += overwrite_len;
 
@@ -62,9 +62,9 @@ bool brf_shift_queue_compare(brf_shift_queue_t *q, const uint8_t *src,
   assert(src);
 
   // Get read offset for accessing queue buffer
-  const unsigned read_offset = q->read_idx & (q->capacity - 1);
+  const size_t read_offset = q->read_idx & (q->capacity - 1);
   // Find the number of bytes can read before wrap around
-  const unsigned pre_wrap_read_len = MIN(size, q->capacity - read_offset);
+  const size_t pre_wrap_read_len = MIN(size, q->capacity - read_offset);
 
   bool ret = !memcmp(&q->buff[read_offset], src, pre_wrap_read_len);
